FAC-984: Feat - Move SpaceStatus and SpaceStatusTest to users module
