+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/domain/SpaceStatus.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.common.application.domain.space;
+package org.flickit.assessment.users.application.domain;
 
 public enum SpaceStatus {
